+++ src/SampleTest/Linq/EnumerableTest.cs
@@ -1,50 +1,6 @@
 namespace SampleTest.Linq;
 
 public class EnumerableTest {
-	private interface ISample {
-	}
-
-	private class Sample : ISample {
-	}
-
-	[Fact]
-	public void Cast_castする() {
-		// Arrange
-		var values = new object[] {
-			new Sample(),
-			new { },
-		};
-
-		// Act
-		var actual = values.Cast<ISample>();
-
-		// Assert
-		// 1つ目はcastできるので取得できる
-		Assert.IsType<Sample>(actual.First());
-
-		// 2つ目を取得しようとするとcastできないので例外が発生する
-		Assert.Throws<InvalidCastException>(() => {
-			actual.Skip(1).First();
-		});
-	}
-
-	[Fact]
-	public void OfType_castできる要素にフィルターする() {
-		// Arrange
-		var values = new object[] {
-			new Sample(),
-			new { },
-		};
-
-		// Act
-		var actual = values.OfType<ISample>();
-
-		// Assert
-		// 変換できる（castできる）要素のみにフィルターされる
-		var sample = Assert.Single(actual);
-		Assert.IsType<Sample>(sample);
-	}
-
 	[Fact]
 	public void Min_非nullの空のシーケンスで呼び出すとInvalidOperationException() {
 		// 非nullの場合は例外
